+++ rustls/src/kx_group/mod.rs
@@ -18,7 +18,7 @@ pub use kem::{MLKEM768, MLKEM1024, X25519MLKEM768};
 /// If the `prefer-post-quantum` feature is enabled, X25519MLKEM768 will
 /// be the first group offered, otherwise it will be the last.
 pub static DEFAULT_KX_GROUPS: &[&dyn SupportedKxGroup] = &[
-	//X25519MLKEM768,
+	X25519MLKEM768,
 	//#[cfg(not(feature = "fips"))]
 	X25519,
 	MLKEM768,